Fix memory leaks and add fallback options
There are a few things going on in this commit. I think it addresses all of Dan's comments.

This commit introduces a command line option called "-fallback", which can currently be set either to `cpu` or `terminate`.
This command line option determines what will happen when gpu memory allocation fails due to insufficient memory.
The terminate option will just throw an error and end execution as soon as a single failed allocation happens. The cpu option will just do the computation on the CPU.
A switch statement that behaves differently based on the value of this option has been added to each operation in the Factorize and Solve methods. This makes it so failed allocations can now be somewhat flexibly handled.

Other things this commit does:
- fixes the memory leaks that were present in the factorize operations.
- replaces all CPU BLAS calls with invocations of a zero-argument lambda expression to make for simpler code